requirements.txt
ffmpeg-python
requests
openai-whisper
selenium